refactor(guard-review): format contract helpers
